Rename api -> controller
