+++ cmd/microservices/profile/main.go
@@ -4,8 +4,6 @@ import (
 	"net/http"
 
 	api "github.com/go-park-mail-ru/2024_2_TeamOn_Patreon/cmd/microservices/profile/api"
-	//behavior "github.com/go-park-mail-ru/2024_2_TeamOn_Patreon/internal/auth/service"
-	//repositories "github.com/go-park-mail-ru/2024_2_TeamOn_Patreon/internal/auth/repository/repositories"
 	logger "github.com/go-park-mail-ru/2024_2_TeamOn_Patreon/internal/pkg/logger"
 	middlewares "github.com/go-park-mail-ru/2024_2_TeamOn_Patreon/internal/pkg/middlewares"
 )
@@ -18,17 +16,12 @@ func main() {
 	// logger
 	logger.New()
 
-	// repository
-	//rep := repositories.New()
-
 	// routers
 	router := api.NewRouter()
 
 	// регистрируем middlewares
-	//commonHandler := middlewares.CreateMiddlewareWithCommonRepository(rep, behavior.New)
 	router.Use(middlewares.Logging)     // 1
 	router.Use(middlewares.HandlerAuth) // 2 только для ручек, где требуется аутентификация
-	//router.Use(commonHandler)           // 3
 
 	// run server
 	logger.StandardInfo("Starting server at: 8082", op)
