+++ internal/profile/controller/profile_get.go
@@ -8,7 +8,7 @@ import (
 	global "github.com/go-park-mail-ru/2024_2_TeamOn_Patreon/internal/pkg/global"
 	"github.com/go-park-mail-ru/2024_2_TeamOn_Patreon/internal/pkg/logger"
 	bModels "github.com/go-park-mail-ru/2024_2_TeamOn_Patreon/internal/pkg/service/models"
-	models "github.com/go-park-mail-ru/2024_2_TeamOn_Patreon/internal/profile/api/models"
+	models "github.com/go-park-mail-ru/2024_2_TeamOn_Patreon/internal/profile/controller/models"
 	repModel "github.com/go-park-mail-ru/2024_2_TeamOn_Patreon/internal/profile/repository/models"
 	repository "github.com/go-park-mail-ru/2024_2_TeamOn_Patreon/internal/profile/repository/repositories"
 )
